added a gameover procdure
This procedure you will call if the answer was wrong or all questions are answered, it will print a message with your score then wait for you to press a key and exit the program. It waits for you to press a key so it doesn't print your score then exit immediatly after so you can read it.